WIP - split files, pytest fixtures
