+++ smart_control/dataset/conftest.py
@@ -1,0 +1,98 @@
+"""Tests for the Smart Buildings Dataset.
+
+Includes high fidelity tests to download the dataset and verify its structure.
+
+It takes around two minutes to download and unzip the data, so we are skipping
+dataset tests by default, to keep the build fast. But you can trigger a download
+by setting the `TEST_DATASET_DOWNLOAD` environment variable to 'true'.
+
+Downloaded data will not get cleared after tests run, so we can use it in
+subsequent test runs without needing to re-download. This allows developers to
+run dataset tests fairly quickly on their local machines. When the dataset
+already exists locally, the tests take around five seconds.
+
+The dataset tests will be run if the data is being downloaded, or if there is
+existing local data.
+
+Downloaded data will not get cleared by default before tests run, but you can
+force a clean up and fresh download by setting the `CLEAR_TEST_DATASET_DOWNLOAD`
+environment variable to 'true'.
+
+These real tests are meant to be run periodically, for example once per day.
+"""
+
+import os
+import shutil
+
+from dotenv import load_dotenv
+import pytest
+
+from smart_control.dataset.dataset import BuildingDataset
+from smart_control.dataset.dataset import DATA_DIR
+from smart_control.dataset.dataset_partition import BuildingDatasetPartition
+
+load_dotenv()
+
+# whether or not to download the dataset:
+TEST_DATASET_DOWNLOAD = bool(os.getenv('TEST_DATASET_DOWNLOAD', default='false').lower() == 'true')  # pylint: disable=line-too-long
+# whether or not to delete existing local data before downloading:
+CLEAR_TEST_DATASET_DOWNLOAD = bool(os.getenv('CLEAR_TEST_DATASET_DOWNLOAD', default='false').lower() == 'true')  # pylint: disable=line-too-long
+
+DATASET_DIRPATH = os.path.join(DATA_DIR, 'sb1')
+ZIP_FILEPATH = os.path.join(DATA_DIR, 'sb1.zip')
+
+# whether or not to run dataset tests:
+TEST_DATASET = bool(TEST_DATASET_DOWNLOAD or os.path.isdir(DATASET_DIRPATH))
+SKIP_REASON = 'Skip large download by default.'
+
+
+def cleanup_files():
+  print('Deleting dataset files...')
+
+  if os.path.isfile(ZIP_FILEPATH):
+    os.remove(ZIP_FILEPATH)
+
+  if os.path.isdir(DATASET_DIRPATH):
+    shutil.rmtree(DATASET_DIRPATH)
+
+
+# PYTEST MODULE LEVEL FIXTURES
+
+
+@pytest.fixture(scope='module')
+def dataset():
+  if TEST_DATASET_DOWNLOAD and CLEAR_TEST_DATASET_DOWNLOAD:
+    cleanup_files()
+
+  print('Downloading and initializing the dataset...')
+  return BuildingDataset(building_id='sb1', download=TEST_DATASET_DOWNLOAD)
+
+
+@pytest.fixture(scope='module')
+def partition(dataset):  # pylint: disable=redefined-outer-name
+  return BuildingDatasetPartition(
+      building_dataset=dataset, partition_id='2022_a'
+  )
+
+
+# SHIMS TO GET PYTEST MODULE FIXTURES TO WORK WITH UNITTEST CLASSES :-)
+
+
+@pytest.fixture(scope='class')
+def inject_dataset(request, dataset):  # pylint: disable=redefined-outer-name
+  """
+  A class-scoped fixture that takes the result of the module-scoped
+  'dataset' fixture and injects it into the test class as 'cls.ds'.
+  """
+  if request.cls:
+    request.cls.ds = dataset
+
+
+@pytest.fixture(scope='class')
+def inject_partition(request, partition):  # pylint: disable=redefined-outer-name, line-too-long
+  """
+  A class-scoped fixture that takes the result of the module-scoped
+  'partition' fixture and injects it into the test class as 'cls.partition'.
+  """
+  if request.cls:
+    request.cls.partition = partition
